requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
loguru>=0.7.0
PyQt6>=6.4.0
PyQt5>=5.15.0
opencv-python>=4.8.0.74
absl_py==2.1.0
accelerate==0.30.1
acme.hello==0.1
colorama==0.4.6
datasets==3.1.0
deepspeed
demjson3==3.0.6
dm_env==1.6
einops==0.8.1
#grpc_tools==1.0.0
grpcio-tools==1.71.0
gym==0.26.2
icecream==2.1.4
ipython==8.12.3
jax==0.6.0
json5==0.10.0
jsonschema==4.23.0
langchain==0.1.12
matplotlib==3.7.4
numpy==1.24.3
openai==1.77.0
packaging==25.0
peft==0.12.0
portpicker==1.6.0
pygame==2.6.1
python_Levenshtein==0.27.1
PyYAML==6.0.2
qwen_agent==0.0.16
qwen_vl_utils==0.0.11
Requests==2.32.3
setuptools==75.1.0
tensorflow==2.19.0
tiktoken==0.7.0
torchvision==0.20.1
tqdm==4.66.3
transformers_stream_generator==0.0.4
trl==0.9.6
ultralytics==8.3.129
vllm==0.7.1
yacs==0.1.8
tf_keras==2.19.0
flash_attn==2.7.4.post1
aiohttp==3.8.5
aiortc==1.5.0
av==10.0.0
